refactor(host): move nats up a level, document
