+++ Cargo.toml
@@ -140,6 +140,7 @@ regex = { workspace = true, optional = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread", "signal"] }
 tracing = { workspace = true } # TODO: revisit the 'release_max_level_info' feature https://github.com/wasmCloud/wasmCloud/issues/468
 tracing-subscriber = { workspace = true, optional = true }
+url = { workspace = true }
 wascap = { workspace = true, optional = true }
 wasmcloud-core = { workspace = true, features = ["otel"], optional = true }
 wasmcloud-host = { workspace = true, optional = true }
