+++ crates/host/src/nats/provider.rs
