+++ crates/host/src/nats/secrets.rs
