+++ crates/host/src/nats/store.rs
